[Maintenance] Move stubbed http client to behat namespace
